% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{Aggsim_infxn_2_death}
\alias{Aggsim_infxn_2_death}
\title{Simulate Aggregate Expected Deaths}
\usage{
Aggsim_infxn_2_death(
  fatalitydata,
  infections,
  m_od = 18.8,
  s_od = 0.45,
  min_day = 1,
  curr_day,
  level,
  simulate_seroprevalence = TRUE,
  specificity,
  sensitivity,
  popN,
  sero_delay_rate
)
}
\arguments{
\item{fatalitydata}{dataframe; The column names: strata, ifr, and pa correspond to (patient) strata, infection-fatality ratio, and the attack rate, respectively.}

\item{infections}{integer vector; The infections for each day up to the current day.}

\item{m_od}{double; The mean of the onset of infection to death (gamma distribution).}

\item{s_od}{double; The coefficient of variation of the onset of infection to death (gamma distribution).}

\item{min_day}{numeric; First day epidemic was observed.}

\item{curr_day}{numeric; Current day of epidemic (considered up to but not including this day).}

\item{level}{character; Must either be "Time-Series" or "Cumulative", indicating whether daily death counts or cumulative deaths to the current day should be returned, respectively.}

\item{simulate_seroprevalence}{logical; Whether or not seroprevalence data should also be simulated}

\item{specificity}{double; Specificity of the Seroprevalence Study (only considered if simulate_seroprevalence is set to TRUE)}

\item{sensitivity}{double; Sensitivity of the Seroprevalence Study (only considered if simulate_seroprevalence is set to TRUE)}

\item{popN}{double; Population Size (only considered if simulate_seroprevalence is set to TRUE)}

\item{sero_delay_rate}{double; Rate of time from infection to seroconversion, assumed to be exponentially distributed (only considered if simulate_seroprevalence is set to TRUE)}

\item{m_or}{double; The mean of the onset of infection to recovery (gamma distribution).}

\item{s_or}{double; The coefficient of variation of the onset of infection to recovery (gamma distribution).}
}
\description{
Simulate Aggregate Expected Deaths
}
