% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{COVIDCurve}
\alias{COVIDCurve}
\title{COVIDCurve: Simple Inference of IFRs and I0 for Aggregate COVID Death Counts}
\description{
Assumes an incidence curve can be mapped on to a onset-to-death curve in order to infer
the IFR or I0 of an epidemic.
}
