% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plot.R
\name{draw_posterior_RGobserved_seroprevalences}
\alias{draw_posterior_RGobserved_seroprevalences}
\title{Draw the Putative Observed Seroprevalnces using the Rogan-Gladen Estimator}
\usage{
draw_posterior_RGobserved_seroprevalences(
  IFRmodel_inf,
  whichrung = "rung1",
  dwnsmpl,
  by_chain = TRUE
)
}
\arguments{
\item{IFRmodel_inf}{R6 class; The result of the IFR Model MCMC run along with the model input.}

\item{whichrung}{character; Specify which rung to sample from (default is rung1)}

\item{dwnsmpl}{integer; Number of posterior results to draw -- weighted by posterior prob}

\item{by_chain}{logical; Whether or not credible intervals should be reported with respect to individual chains (TRUE) or not.}
}
\description{
Draw the Putative Observed Seroprevalnces using the Rogan-Gladen Estimator
}
\details{
Given sampling iterations with posterior-log-likes greater than or equal to a specific threshold, posterior results for the linear spline are generated. Assumed that the spline was fit in "un-transformed" space
}
