% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plot.R
\name{get_gelman_rubin_diagnostic}
\alias{get_gelman_rubin_diagnostic}
\title{Get Gelman Diagnostic for Age-Specific Model}
\usage{
get_gelman_rubin_diagnostic(IFRmodel_inf)
}
\arguments{
\item{IFRmodel_inf}{R6 class; The result of the IFR Model MCMC run along with the model input.}
}
\description{
Get Gelman Diagnostic for Age-Specific Model
}
\details{
Using the \link[coda]{gelman.diag} function to assess convergence
}
