% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plot.R
\name{get_globalIFR_cred_intervals}
\alias{get_globalIFR_cred_intervals}
\title{Get the Global IFR Credible Intervals, or IFR Weighted by Demography}
\usage{
get_globalIFR_cred_intervals(
  IFRmodel_inf,
  whichrung = "rung1",
  by_chain = TRUE
)
}
\arguments{
\item{IFRmodel_inf}{R6 class; The result of the IFR Model MCMC run along with the model input.}

\item{whichrung}{character; Specify which rung to sample from (default is rung1)}

\item{by_chain}{logical; Whether or not credible intervals should be reported with respect to individual chains (TRUE) or not.}
}
\description{
Get the Global IFR Credible Intervals, or IFR Weighted by Demography
}
