% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plot.R
\name{get_overall_IFR_cred_intervals}
\alias{get_overall_IFR_cred_intervals}
\title{Get the overall IFR Weighted by Demography}
\usage{
get_overall_IFR_cred_intervals(
  IFRmodel_inf,
  whichstandard = "pop",
  whichrung = "rung1",
  by_chain = TRUE
)
}
\arguments{
\item{IFRmodel_inf}{R6 class; The result of the IFR Model MCMC run along with the model input.}

\item{whichstandard}{character; Whether the population demography weighted (pop) or the attack-rate weighted population (arpop) standardization should be applied}

\item{whichrung}{character; Specify which rung to sample from (default is rung1)}

\item{by_chain}{logical; Whether or not credible intervals should be reported with respect to individual chains (TRUE) or not.}
}
\description{
Get the overall IFR Weighted by Demography
}
