% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelistsim.R
\name{LineListsim_infxn_2_death}
\alias{LineListsim_infxn_2_death}
\title{Simulate Aggregate Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve}
\usage{
LineListsim_infxn_2_death(
  casefat,
  I0,
  r,
  m_od = 18.8,
  s_od = 0.45,
  m_or = 24.7,
  s_or = 0.45,
  hospprob,
  min_day = 1,
  curr_day,
  expgrowth,
  infections
)
}
\arguments{
\item{casefat}{dataframe; column names age, ifr, and pa that correspond to age-band, infection-fatality ratio, and the attack rate in given age band, respectively}

\item{I0}{integer; number of infected individuals at the beginning of the epidemic for exponential growth}

\item{r}{double; the expontential growth rate of the epidemic}

\item{m_od}{double; the mean of the onset of infection to death (gamma distribution)}

\item{s_od}{double; the coefficient of variation of the onset of infection to death (gamma distribution)}

\item{m_or}{double; the mean of the onset of infection to recovery (gamma distribution)}

\item{s_or}{double; the coefficient of variation of the onset of infection to recovery (gamma distribution)}

\item{hospprob}{numeric; probability of hospitalization given infection}

\item{min_day}{numeric; first day epidemic was observed}

\item{curr_day}{numeric; current day of epidemic, considered up to but not including this day}

\item{expgrowth}{logical; Simulate exponential growth or not}

\item{infections}{integer vector; The infections for each day}
}
\description{
Simulate Aggregate Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve
}
