% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{make_IFRmodel_age}
\alias{make_IFRmodel_age}
\title{Make Model Inference Object-Oriented Class}
\description{
A simple R6 class with only public features. Purpose is to provide a model framework
for inference of IFRs from Age-Specific Death Data.
}
