% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{make_IFRmodel_agg}
\alias{make_IFRmodel_agg}
\title{Make Model Inference Object-Oriented Class}
\description{
A simple R6 class with only public features. Purpose is to provide a model framework
for inference of IFRs and the incidence curve from Aggregate Death Data.
}
\section{Public Variables}{

\tabular{ll}{
\strong{Variable} \tab \strong{Return} \cr
\code{data} \tab  Input data for inference \cr
}
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{data} \tab dataframe \tab Dataframe with .... \cr
}
}

