% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOO.R
\name{make_modinf_linelist}
\alias{make_modinf_linelist}
\title{Object-Oriented Class for Model Inference with LineList}
\description{
A simple R6 class with only public features.
}
\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{data} \tab dataframe \tab Dataframe with .... \cr
}
}

\section{Public Variables}{

\tabular{ll}{
\strong{Variable} \tab \strong{Return} \cr
\code{data} \tab  Input data for ifnerence \cr
}
}

