% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggcpp2strings.R
\name{make_user_Agg_loglike}
\alias{make_user_Agg_loglike}
\title{Liftover User Param Df to Likelihood Function for Dr. Jacoby}
\usage{
make_user_Agg_loglike(modinf, reparamIFR)
}
\description{
Liftover User Param Df to Likelihood Function for Dr. Jacoby
}
