% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggcpp2strings_noreparam.R
\name{make_user_Agg_loglike_noreparam}
\alias{make_user_Agg_loglike_noreparam}
\title{Liftover User Param Df to Likelihood Function for Dr. Jacoby}
\usage{
make_user_Agg_loglike_noreparam(modinf)
}
\description{
Liftover User Param Df to Likelihood Function for Dr. Jacoby
}
