% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggcpp2strings.R
\name{make_user_Agg_logprior}
\alias{make_user_Agg_logprior}
\title{Liftover User Param Df to Prior Function for Dr. Jacoby}
\usage{
make_user_Agg_logprior(modinf, reparamIFR)
}
\description{
Liftover User Param Df to Prior Function for Dr. Jacoby
}
