% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineListcpp2strings_reparam.R
\name{make_user_LineList_loglike_reparam}
\alias{make_user_LineList_loglike_reparam}
\title{Liftover User Param Df to Likelihood Function for Dr. Jacoby}
\usage{
make_user_LineList_loglike_reparam(modinf)
}
\description{
Liftover User Param Df to Likelihood Function for Dr. Jacoby
}
