% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineListcpp2strings_reparam.R
\name{make_user_LineList_logprior_reparam}
\alias{make_user_LineList_logprior_reparam}
\title{Liftover User Param Df to Prior Function for Dr. Jacoby}
\usage{
make_user_LineList_logprior_reparam(modinf)
}
\description{
Liftover User Param Df to Prior Function for Dr. Jacoby
}
