% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_IFRmodel_age}
\alias{run_IFRmodel_age}
\title{Run Aggregate Model}
\usage{
run_IFRmodel_age(
  IFRmodel,
  binomial_likelihood = TRUE,
  reparamIFR = TRUE,
  reparamInfxn = TRUE,
  reparamKnots = TRUE,
  thinning = 0,
  burnin = 1000,
  samples = 10000,
  rungs = 1,
  chains = 5,
  coupling_on = TRUE,
  GTI_pow = 1,
  beta_manual = NULL,
  cluster = NULL,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{IFRmodel}{R6 class; Internal model object for COVIDCurve}

\item{binomial_likelihood}{logical; Whether the binomial or the logit likelihood should be used}

\item{reparamIFR}{logical; Whether IFRs should be reparameterized or inferred separately}

\item{reparamInfxn}{logical; Whether infection curve (i.e. the  y-coordinates infection spline) should be reparameterized or inferred separately}

\item{reparamKnots}{logical; Whether infection knots (i.e. the x-coordinates of the infection spline) should be reparameterized or inferred separately}

\item{thinning}{integer; The regular sequence to count by to thin MCMC posterior chain (iterations are kept as: \code{seq(from = thinning, to = (burnin+samples), by = thinning)}).}

\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{rungs}{the number of temperature rungs used in Metropolis coupling (see
\code{coupling_on}).}

\item{chains}{the number of independent replicates of the MCMC chain to run.
If a \code{cluster} object is defined then these chains are run in
parallel, otherwise they are run in serial.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs.}

\item{GTI_pow}{values in the temperature ladder are raised to this power.}

\item{beta_manual}{option to manually define temperature ladder. These values
are raised to the power \code{GTI_pow}, hence you should use \code{GTI_code
= 1} to fix powers exactly. If \code{NULL} then an equal spacing of length
\code{rungs} is used between 0 and 1.}

\item{cluster}{option to pass in a cluster environment, allowing chains to be
run in parallel (see package "parallel").}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
Run Aggregate Model
}
\details{
Wraps the Metropolic-Coupled MCMC Framework from Dr. Jacoby
}
