% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_IFRmodel_agg}
\alias{run_IFRmodel_agg}
\title{Run Aggregate Model}
\usage{
run_IFRmodel_agg(
  IFRmodel,
  reparamIFR = TRUE,
  reparamInfxn = TRUE,
  reparamKnots = TRUE,
  reparamSpec = TRUE,
  burnin = 1000,
  samples = 1000,
  chains = 3,
  rungs = 1,
  GTI_pow = 3,
  coupling_on = TRUE,
  cluster = NULL,
  pb_markdown = FALSE,
  silent = TRUE
)
}
\arguments{
\item{IFRmodel}{R6 class; Internal model object for COVIDCurve}

\item{reparamIFR}{logical; Whether IFRs should be reparameterized or inferred seperately}

\item{reparamInfxn}{logical; Whether infection curve (i.e. the  y-coordinates infection spline) should be reparameterized or inferred seperately}

\item{reparamKnots}{logical; Whether infection knots (i.e. the x-coordinates of the infection spline) should be reparameterized or inferred seperately}

\item{reparamSpec}{logical; Whether Specificity should be reparameterized or inferred seperately. It is inferred relative to the MaxMa, and thus the user must also reparameterize the IFRs to consider this option}

\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{chains}{the number of independent replicates of the MCMC chain to run.
If a \code{cluster} object is defined then these chains are run in
parallel, otherwise they are run in serial.}

\item{rungs}{the number of temperature rungs used in Metropolis coupling (see
\code{coupling_on}).}

\item{GTI_pow}{the power used in the generalised thermodynamic integration 
method.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs.}

\item{cluster}{option to pass in a cluster environment, allowing chains to be
run in parallel (see package "parallel").}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
Run Aggregate Model
}
\details{
Wraps the Metropolic-Coupled MCMC Framework from Dr. Jacoby

Within the model fitting, the mean delay of seroconversion is always considered as a scale of the mean delay of the onset-to-death for the gamma distribution
}
