% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_modinf_linelist}
\alias{run_modinf_linelist}
\title{Run Line-List Model}
\usage{
run_modinf_linelist(
  modinf,
  reparamIFR = T,
  burnin = 1000,
  samples = 1000,
  chains = 3,
  rungs = 1,
  GTI_pow = 3,
  coupling_on = T,
  pb_markdown = F,
  silent = T
)
}
\arguments{
\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{chains}{the number of independent replicates of the MCMC chain to run.
If a \code{cluster} object is defined then these chains are run in
parallel, otherwise they are run in serial.}

\item{rungs}{the number of temperature rungs used in Metropolis coupling (see
\code{coupling_on}).}

\item{GTI_pow}{the power used in the generalised thermodynamic integration 
method.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
Run Line-List Model
}
\details{
Wraps the Metropolic-Coupled MCMC Framework from Dr. Jacoby
}
