% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{sim_infxn_2_death}
\alias{sim_infxn_2_death}
\title{Simulate Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve}
\usage{
sim_infxn_2_death(
  casefat,
  I0,
  r,
  m_od = 18.8,
  s_od = 0.45,
  min_day = 1,
  curr_day,
  level,
  expgrowth,
  infections
)
}
\arguments{
\item{casefat}{dataframe; column names age, cfr, and pa that correspond to age-band, case-fatality ratio, and the attack rate in given age band, respectively}

\item{I0}{integer; number of infected individuals at the beginning of the epidemic}

\item{r}{double; the expontential growth rate of the epidemic}

\item{m_od}{double; the mean of the onset of infection to death (gamma distribution)}

\item{s_od}{double; the coefficient of variation of the onset of infection to death (gamma distribution)}

\item{min_day}{numeric; first day epidemic was observed}

\item{curr_day}{numeric; current day of epidemic, considered up to but not including this day}

\item{level}{character; Either "Cumulative" or "Time-Series" if expected deaths should be returned as a total or for each day, respectively}

\item{expgrowth}{logical; Simulate exponential growth or not}

\item{infections}{integer vector; The infections for each day}
}
\description{
Simulate Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve
}
