% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{sim_infxn_2_death_timeseries}
\alias{sim_infxn_2_death_timeseries}
\title{Simulate Expected Deaths in a Time-Series}
\usage{
sim_infxn_2_death_timeseries(
  casefat,
  I0,
  r,
  curr_day,
  m_od = 18.8,
  s_od = 0.45
)
}
\arguments{
\item{casefat}{dataframe; column names age, cfr, and pa that correspond to age-band, case-fatality ratio, and the attack rate in given age band, respectively}

\item{I0}{integer; number of infected individuals at the beginning of the epidemic}

\item{r}{double; the expontential growth rate of the epidemic}

\item{curr_day}{numeric; current day of epidemic}

\item{m_od}{double; the mean of the onset of infection to death (gamma distribution)}

\item{s_od}{double; the coefficient of variation of the onset of infection to death (gamma distribution)}
}
\description{
Simulate Expected Deaths in a Time-Series
}
