% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{sim_seroprev}
\alias{sim_seroprev}
\title{Simulate Seroprevalence Study}
\usage{
sim_seroprev(
  infxns,
  specificity,
  sensitivity,
  sero_delay_rate,
  popN,
  fatalitydata,
  min_day,
  curr_day
)
}
\arguments{
\item{infxns}{numeric vector; the expected number of infections from the infection curve from the poisson draw
internal function, not exported}

\item{specificity}{double; Specificity of the Seroprevalence Study (only considered if simulate_seroprevalence is set to TRUE)}

\item{sensitivity}{double; Sensitivity of the Seroprevalence Study (only considered if simulate_seroprevalence is set to TRUE)}

\item{sero_delay_rate}{double; Rate of time from infection to seroconversion, assumed to be exponentially distributed (only considered if simulate_seroprevalence is set to TRUE)}

\item{popN}{double; Population Size (only considered if simulate_seroprevalence is set to TRUE)}

\item{fatalitydata}{dataframe; The column names: strata, ifr, and pa correspond to (patient) strata, infection-fatality ratio, and the attack rate, respectively.}

\item{min_day}{numeric; First day epidemic was observed.}

\item{curr_day}{numeric; Current day of epidemic (considered up to but not including this day).}
}
\description{
Simulate Seroprevalence Study
}
