% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{wrap_drjacoby_mcmc}
\alias{wrap_drjacoby_mcmc}
\title{Wrap the Metropolic-Coupled MCMC Framework from Dr. Jacoby}
\usage{
wrap_drjacoby_mcmc(
  data,
  df_params,
  misc,
  LogLike,
  LogPrior,
  level,
  reparameterization,
  burnin = 1000,
  samples = 1000,
  chains = 3,
  rungs = 1,
  GTI_power = 3,
  coupling_on = T,
  pb_markdown = F,
  silent = T
)
}
\arguments{
\item{data}{a vector of data values. When using C++ versions of the
likelihood and/or prior these values are treated internally as doubles, so
while integer and boolean values can be used, keep in mind that these will
be recast as doubles in the likelihood (i.e. \code{TRUE = 1.0}).}

\item{df_params}{a dataframe of parameters. Must contain the following
elements:
\itemize{
  \item \code{name} - the parameter name.
  \item \code{min} - the minimum value of the parameter. \code{-Inf} is
  allowed.
  \item \code{max} - the maximum value of the parameter. \code{Inf} is
  allowed.
  \item \code{init} - the initial value of the parameter.
}}

\item{misc}{list; Required list object that must contain the following items:
elements:
\itemize{
\item \code{min_day} - the presumed minimum day of the epidemic (required only for time-series likelihoods)
\item \code{curr_day} - the presumed minimum day of the epidemic.
\item \code{pa} - A vector of the attack rates for the given age bands
}}

\item{LogLike}{global function;}

\item{LogPrior}{global function;}

\item{level}{character; Must be either Time-Series or Cumulative}

\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{chains}{the number of independent replicates of the MCMC chain to run.
If a \code{cluster} object is defined then these chains are run in
parallel, otherwise they are run in serial.}

\item{rungs}{the number of temperature rungs used in Metropolis coupling (see
\code{coupling_on}).}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
Wrap the Metropolic-Coupled MCMC Framework from Dr. Jacoby
}
