% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{Aggsim_infxn_2_death}
\alias{Aggsim_infxn_2_death}
\title{Simulate Aggregate Expected Deaths}
\usage{
Aggsim_infxn_2_death(
  fatalitydata,
  infections,
  m_od = 14.26,
  s_od = 0.79,
  curr_day,
  spec,
  sens,
  demog,
  sero_delay_rate,
  simulate_seroreversion,
  sero_rev_shape = NULL,
  sero_rev_scale = NULL,
  smplfrac = 1,
  return_linelist = FALSE
)
}
\arguments{
\item{fatalitydata}{dataframe; The strata-specific fatalities to simulate given a probability of infection and a noise effect. The column names: strata, ifr, rho, and Ne correspond to (patient) strata, infection-fatality ratio, the probability of infection (i.e. a probalistic attack rate), and a noise effect, respectively.}

\item{infections}{integer vector; The infections for each day up to the current day.}

\item{m_od}{double; The mean of the onset of infection to death (gamma distribution).}

\item{s_od}{double; The coefficient of variation of the onset of infection to death (gamma distribution).}

\item{curr_day}{numeric; Current day of epidemic (considered up to but not including this day).}

\item{spec}{double; Specificity of the Seroprevalence Study (only considered if simulate_seroprevalence is set to TRUE)}

\item{sens}{double; Sensitivity of the Seroprevalence Study (only considered if simulate_seroprevalence is set to TRUE)}

\item{demog}{dataframe; Strata-specific population (demographic) counts. The columns names strata and popN correspond to (patient) strata and the number of individuals within that strata. The demography strata must match the fatalitydata strata. Only considered if \code{simulate_seroprevalence = TRUE}}

\item{sero_delay_rate}{double; Rate of time from infection to seroconversion, assumed to be exponentially distributed (only considered if simulate_seroprevalence is set to TRUE)}

\item{simulate_seroreversion}{logical; Whether seroreversion (due to waning of antibodies) should be simulated or not}

\item{sero_rev_shape}{double; The shape parameter of the Weibull seroreversion distribution}

\item{sero_rev_scale}{double; The scale parameter of the Weibull seroreversion distribution}

\item{smplfrac}{numeric; Sampling fraction for the observed seroprevalence study (assumed to be a simple random sample of all infected)}

\item{return_linelist}{logical; Whether or not the linelist that was used to create the observed marginal data should be returned. N.B. the linelist can be quite large/burdensome for memory depending on population size and number of days considered.}

\item{level}{character; Must either be "Time-Series" or "Cumulative", indicating whether daily death counts or cumulative deaths to the current day should be returned, respectively.}
}
\description{
Simulate Aggregate Expected Deaths
}
