% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plot.R
\name{get_cred_intervals}
\alias{get_cred_intervals}
\title{Get Credible Intervals for Parameters from Sampling Iterations}
\usage{
get_cred_intervals(IFRmodel_inf, what, whichrung = "rung1", by_chain = TRUE)
}
\arguments{
\item{IFRmodel_inf}{R6 class; The result of the IFR Model MCMC run along with the model input.}

\item{what}{character; Specify which parameters you want: "IFRparams", "Infxnparams", or "Seroparams"}

\item{whichrung}{character; Specify which rung to sample from (default is rung1)}

\item{by_chain}{logical; Whether or not credible intervals should be reported with respect to individual chains (TRUE) or not.}
}
\description{
Get Credible Intervals for Parameters from Sampling Iterations
}
