% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{LineListsim_infxn_2_death}
\alias{LineListsim_infxn_2_death}
\title{Simulate Line-List Outcomes}
\usage{
LineListsim_infxn_2_death(
  casefat,
  infections,
  m_od = 18.8,
  s_od = 0.45,
  m_or = 24.7,
  s_or = 0.45,
  hospprob,
  min_day = 1,
  curr_day,
  simulate_seroprevalence = FALSE
)
}
\arguments{
\item{casefat}{dataframe; The column names: age, ifr, and pa correspond to age-band, infection-fatality ratio, and the attack rate, respectively.}

\item{infections}{integer vector; The infections for each day up to the current day (must be supplied if \param{expgrowth} is false).}

\item{m_od}{double; The mean of the onset of infection to death (gamma distribution).}

\item{s_od}{double; The coefficient of variation of the onset of infection to death (gamma distribution).}

\item{m_or}{double; The mean of the onset of infection to recovery (gamma distribution).}

\item{s_or}{double; The coefficient of variation of the onset of infection to recovery (gamma distribution).}

\item{hospprob}{numeric; Probability of hospitalization given infection.}

\item{min_day}{numeric; First day epidemic was observed.}

\item{curr_day}{numeric; Current day of epidemic (considered up to but not including this day).}

\item{simulate_seroprevalence}{logical; Boolean indicating whether or not the seroprevalence should also be simulated}
}
\description{
Simulate Line-List Outcomes
}
