% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plot.R
\name{posterior_check_infxns_to_death}
\alias{posterior_check_infxns_to_death}
\title{Posterior Check for Infections to Death}
\usage{
posterior_check_infxns_to_death(
  IFRmodel_inf,
  whichrung = "rung1",
  CIquant,
  by_chain = FALSE
)
}
\arguments{
\item{IFRmodel_inf}{R6 class; The result of the IFR Model MCMC run along with the model input.}

\item{whichrung}{character; Specify which rung to sample from (default is rung1)}

\item{by_chain}{logical; Whether or not credible intervals should be reported with respect to individual chains (TRUE) or not.}
}
\description{
Posterior Check for Infections to Death
}
\details{
Given sampling iterations with posterior-log-likes greater than or equal to a specific threshold, posterior results for the linear spline are generated. Assumed that the spline was fit in "un-transformed" space
}
