% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{sim_seroprev}
\alias{sim_seroprev}
\title{Simulate Seroprevalence Study}
\usage{
sim_seroprev(
  infxns,
  day,
  sero_spec,
  sero_sens,
  popN,
  sero_delay_rate,
  casefat,
  min_day,
  curr_day
)
}
\arguments{
\item{sero_spec}{double; (only considered if simulate_seroprevalence is set to TRUE)}

\item{sero_sens}{double; (only considered if simulate_seroprevalence is set to TRUE)}

\item{popN}{double; (only considered if simulate_seroprevalence is set to TRUE)}

\item{sero_delay_rate}{double; (only considered if simulate_seroprevalence is set to TRUE)
internal function, not exported}

\item{casefat}{dataframe; The column names: age, ifr, and pa correspond to age-band, infection-fatality ratio, and the attack rate, respectively.}

\item{min_day}{numeric; First day epidemic was observed.}

\item{curr_day}{numeric; Current day of epidemic (considered up to but not including this day).}
}
\description{
Simulate Seroprevalence Study
}
